refactor(chat): extract chat_execution_service — /chat applier + setup (#1483)
Move the /chat business logic out of routers/chat.py into
services/chat_execution_service.py (Invariant #1), decomposing the CC-57
_run_chat_and_finalize monster (AC #2):

- prepare_chat_execution (was _prepare_chat_execution): exec record + subscription
  + collaboration broadcast/activity + session + chat-start activity + user-msg log.
- broadcast_collaboration_event (moved, own set_websocket_manager wired in main.py).
- run_chat_turn (was _run_chat_and_finalize, declared TRANSITIONAL — RD15) split into
  build_chat_payload / _finalize_chat_success / _parse_agent_http_error /
  _finalize_budget_exhausted / _finalize_http_failure / _apply_sub003_autoswitch.
  Each ≤ CC 20. HTTP-free: failure paths raise ChatDispatchError, mapped 1:1 by the
  thin chat_with_agent handler; the lone fastapi touch is the defensive
  `except HTTPException: raise` preserving SUB-003 propagate-unchanged semantics.

Byte-for-byte preserved: MEM-001 runtime-aware prompt, #686 mark-dispatched-before-POST,
UUID-validated claude_session_id, #1332 read-before-close mirroring, #678 partial-
metadata salvage, the full SUB-003 429/auth switch matrix, and the finally slot+idem
release. routers/chat.py: 2756 → 1889 lines; unused imports trimmed. OpenAPI
byte-identical (372 paths).

Tests repointed in-commit: the run_chat_turn char suite + test_chat_admission's direct
finalize/prepare tests (now raise ChatDispatchError, patch _CE collaborators); the
test_chat_dispatched_marker AST/mirror guards now parse chat_execution_service.py.
